replace gallery route w/ photo
