+++ app/photo/[id]/[name]/page.tsx
